student creation upadte student info delete student

+++ src/main/java/mum/edu/cs/cs425/eregister/repository/IstudentRepository.java
@@ -1,11 +1,14 @@
 package mum.edu.cs.cs425.eregister.repository;
 
+import java.util.Optional;
+
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import mum.edu.cs.cs425.eregister.models.Student;
 
 @Repository
-public interface IstudentRepository extends CrudRepository<Student,Long> {
+public interface IstudentRepository extends CrudRepository<Student, Integer> {
+    //Optional<Student> findStudentById(Long id);
     
 }
